# swagger/swagger.yaml
swagger: '2.0'
info:
  title: Gerenciamento de Funcionários API
  description: |-
    Sobre a API:
      - A API de gerenciamento de funcionários permite a criação, atualização, listagem e exclusão de funcionários que estão vinculados a departamentos. 
      - Cada funcionário é associado a um departamento específico e possui informações como nome, email e data de criação.

    Links uteis:
    - [Repositório Cadastra Funcionário API](https://github.com/angelozero/backend)
  version: 1.0.0
  contact:
    email: angelo.akm@gmail.com
basePath: /
tags:
  - name: Funcionários
    description: Operações relacionadas a funcionários
  - name: Departamentos
    description: Operações relacionadas a departamentos
  - name: Satisfação
    description: Operações relacionadas a predição de satisfação de um funcinário

paths:
  /api/funcionarios:
    get:
      tags:
        - Funcionários
      summary: Listando funcionários
      parameters:
        - name: pagina
          in: query
          description: Correspondente à página de pesquisa
          required: false
          type: integer
          default: 1
        - name: total
          in: query
          description: Número de registros por página
          required: false
          type: integer
          default: 5
        - name: nome
          in: query
          description: Nome do funcionário
          required: false
          type: string
        - name: email
          in: query
          description: Email do funcionário
          required: false
          type: string  
        - name: departamento
          in: query
          description: Nome do departamento
          required: false
          type: string     
      responses:
        '200':
          description: Uma lista paginada de funcionários
          schema:
            type: object
            properties:
              page:
                type: integer
                description: Número da página atual
              total_pages:
                type: integer
                description: Número total de páginas
              total_records:
                type: integer
                description: Número total de registros
              results:
                type: array
                items:
                  $ref: '#/definitions/Employee'
  /api/funcionario:
    post:
      tags:
        - Funcionários
      summary: Cria um novo funcionário
      parameters:
        - name: body
          in: body
          description: Dados do novo funcionário
          required: true
          schema:
            $ref: '#/definitions/NewEmployee'
      responses:
        '201':
          description: Funcionário criado com sucesso
          schema:
            $ref: '#/definitions/Employee'
        '400':
          description: |- 
            - Nome não foi enviado ou está vazio
            - Email não foi enviado ou está vazio
            - Email já cadastrado
            - Departamento inválido
            - CEP não foi enviado ou não é valido
        '404':
          description: Departamento não encontrado
  /api/funcionario/{id}:
    get:
      tags:
        - Funcionários
      summary: Detalhes do funcionário
      parameters:
        - name: id
          in: path
          description: ID do funcionário
          required: true
          type: integer
      responses:
        '200':
          description: Detalhes do funcionário
          schema:
            $ref: '#/definitions/Employee'
        '404':
          description: Funcionário não encontrado
    put:
      tags:
        - Funcionários
      summary: Atualiza um funcionário existente
      parameters:
        - name: id
          in: path
          description: ID do funcionário
          required: true
          type: integer
        - name: body
          in: body
          description: Dados para atualizar as informações do funcionário
          required: true
          schema:
            $ref: '#/definitions/UpdateEmployee'
      responses:
        '200':
          description: Funcionário atualizado com sucesso
          schema:
            $ref: '#/definitions/Employee'
        '400':
          description: |-
            - Dados para atualização não enviados
            - Email inválido
            - Email já cadastrado
            - Departamento inválido
            - CEP não foi enviado ou não é valido
        '404':
          description: |-
            - Funcionário ou Departamento não encontrado
            - Endereço vinculado ao funcionário não encontrado
    delete:
      tags:
        - Funcionários
      summary: Exclui um funcionário existente
      parameters:
        - name: id
          in: path
          description: ID do funcionário
          required: true
          type: integer
      responses:
        '200':
          description: Funcionário excluído com sucesso
        '404':
          description: Funcionário não encontrado
  /api/departamentos:
    get:
      tags:
        - Departamentos
      summary: Lista todos os departamentos
      responses:
        '200':
          description: Lista de todos os departamentos
          schema:
            type: array
            items:
              $ref: '#/definitions/Department'
  /api/departamento:
    post:
      tags:
        - Departamentos
      summary: Cria um novo departamento
      parameters:
        - name: body
          in: body
          description: Dados do novo departamento
          required: true
          schema:
            $ref: '#/definitions/NewDepartment'
      responses:
        '201':
          description: Departamento criado com sucesso
          schema:
            $ref: '#/definitions/Department'
        '400':
          description: Requisição inválida, o campo 'name' é obrigatório
        '409':
          description: Conflito, departamento já existe
  /api/departamento/{id}:
    get:
      tags:
        - Departamentos
      summary: Detalhes do departamento
      parameters:
        - name: id
          in: path
          description: ID do departamento
          required: true
          type: integer
      responses:
        '200':
          description: Detalhes do departamento
          schema:
            $ref: '#/definitions/Department'
        '404':
          description: Departamento não encontrado
  /api/predict:
    post:
      tags:
        - Predição
      summary: Consulta uma predição
      parameters:
        - name: body
          in: body
          description: Dados para predição
          required: true
          schema:
            $ref: '#/definitions/Prediction'
      responses:
        '200':
          description: Predição calculada com sucesso
          schema:
            $ref: '#/definitions/PredictionResponse'

definitions:
  Prediction:
    type: object
    properties:
      satisfaction_level:
        type: number
        description: Nível de satisfação do funcionário
      last_evaluation:
        type: number
        description: Avaliação mais recente do funcionário
      number_project:
        type: integer
        description: Número de projetos em que o funcionário está envolvido
      average_montly_hours:
        type: integer
        description: Média de horas mensais trabalhadas pelo funcionário
      time_spend_company:
        type: integer
        description: Tempo que o funcionário passou na empresa (em anos)
      work_accident:
        type: integer
        description: Indica se o funcionário teve um acidente de trabalho (0 = não, 1 = sim)
      promotion_last_5years:
        type: integer
        description: Indica se o funcionário foi promovido nos últimos 5 anos (0 = não, 1 = sim)
      dept:
        type: string
        description: Departamento do funcionário
      salary:
        type: string
        description: Nível salarial do funcionário (ex. 'low', 'medium', 'high')
  
  PredictionResponse:
    type: object
    properties:
      prediction:
        type: number
        description: Resultado da predição

  UpdateAddress:
    type: object
    properties:
      zipcode:
        type: string
        description: CEP do funcionário
      street:
        type: string
        description: Rua do funcionário
      complement:
        type: string
        description: Complemento do endereço do funcionário
      number:
        type: string
        description: Número da residência do funcionário
      neighborhood:
        type: string
        description: Bairro do funcionário

  Address:
    type: object
    properties:
      zipcode:
        type: string
        description: CEP do funcionário
      street:
        type: string
        description: Rua do funcionário
      complement:
        type: string
        description: Complemento do endereço do funcionário
      number:
        type: string
        description: Número da residência do funcionário
      city:
        type: string
        description: Cidade do funcionário
      neighborhood:
        type: string
        description: Bairro do funcionário
      uf:
        type: string
        description: UF do funcionário

  CreateAddress:
    type: object
    properties:
      zipcode:
        type: string
        description: CEP do funcionário
      street:
        type: string
        description: Rua do funcionário
      complement:
        type: string
        description: Complemento do endereço do funcionário
      number:
        type: string
        description: Número da residência do funcionário
      neighborhood:
        type: string
        description: Bairro do funcionário

  Employee:
    type: object
    properties:
      id:
        type: integer
        description: ID do funcionário
      name:
        type: string
        description: Nome do funcionário
      email:
        type: string
        description: Email do funcionário
      department:
        $ref: '#/definitions/Department'
        description: Departamento do funcionário
      address:
        $ref: '#/definitions/Address'
        description: Endereço do funcionário
      date_time_creation:
        type: string
        format: date-time
        description: Data e hora de criação do funcionário
      date_time_updated:
        type: string
        format: date-time
        description: Data e hora da última atualização do funcionário

  NewEmployee:
    type: object
    properties:
      name:
        type: string
        description: Nome do funcionário
      second_name:
        type: string
        description: Segundo nome do funcionário
      email:
        type: string
        description: Email do funcionário
      department_id:
        type: integer
        description: ID do departamento
      address:
        $ref: '#/definitions/CreateAddress'
        description: Endereço do funcionário

  UpdateEmployee:
    type: object
    properties:
      name:
        type: string
        description: Novo nome do funcionário
      second_name:
        type: string
        description: Novo sobrenome do funcionário
      email:
        type: string
        description: Novo email do funcionário
      department_id:
        type: integer
        description: Novo ID do departamento do funcionário
      address:
        $ref: '#/definitions/UpdateAddress'
        description: Endereço do funcionário

  Department:
    type: object
    properties:
      id:
        type: integer
        description: ID do departamento
      name:
        type: string
        description: Nome do departamento

  NewDepartment:
    type: object
    properties:
      name:
        type: string
        description: Nome do departamento